"fixed mysql, set up secret info, checked repositories and models"

+++ src/main/java/com/punkpad/app/repository/ProgressionPatternStepRepository.java
@@ -1,5 +1,6 @@
 package com.punkpad.app.repository;
 
+import com.punkpad.app.model.Chord;
 import com.punkpad.app.model.ProgressionPatternStep;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProgressionPatternStepRepository extends JpaRepository<ProgressionPatternStep, Long> {
+public interface ProgressionPatternStepRepository extends JpaRepository<ProgressionPatternStep, Integer> {
 
     List<ProgressionPatternStep> findByPatternIdOrderByStepOrderAsc(Long patternId);
 }
